Renaming multiSwap -> sell
